+++ api/ngrok/v1alpha1/kubernetesoperator_types.go
@@ -42,6 +42,8 @@ type ngrokAPICommon struct {
 // NOTE: json tags are required.  Any new fields you add must have json tags for the fields to be serialized.
 
 type KubernetesOperatorDeployment struct {
+	// Cluster is the name of the k8s cluster in which the operator is deployed
+	Cluster string `json:"cluster,omitempty"`
 	// Name is the name of the k8s deployment for the operator
 	Name string `json:"name,omitempty"`
 	// The namespace in which the operator is deployed